Push beginning of migrated installer - Parts are tested, however the fully cycle is not
