+++ app/views/install/install_base.html
@@ -1,0 +1,43 @@
+<!DOCTYPE html>
+<html lang="en">
+    <head>
+	{% block head %}
+	{% block scripts %}
+	<meta charset="UTF-8">
+	<script type="text/javascript">var pageRenderStart = new Date().getTime();</script>
+	<link rel="stylesheet" href="../../assets/css/bootstrap.css">
+	<script src="//ajax.googleapis.com/ajax/libs/jquery/1.11.0/jquery.min.js"></script>
+	{% endblock %}
+	<title>PufferPanel Installer</title>
+	{% endblock %}
+    </head>
+    
+    <body>
+	<div class="container">
+	    <div class="alert alert-danger">
+		<strong>WARNING:</strong> Do not run this version on a live environment! There are known security holes that we are working on getting patched. This is extremely beta software and this version is to get the features in place while we work on security enhancements.
+	    </div>
+	    <div class="navbar navbar-default">
+		<div class="navbar-header">
+		    <a class="navbar-brand" href="#">Install PufferPanel</a>
+		</div>
+	    </div>
+	    <div class="col-12">
+		<div class="row">
+		    <div class="col-2"></div>
+		    <div class="col-8">
+			{% block content %}{% endblock %}
+		    </div>
+		</div>
+	    </div>
+	    <div class="footer">
+		<div class="row">
+		    <div class="col-8 ">
+			<p>PufferPanel is licensed under a <a href="https://github.com/PufferPanel/PufferPanel/blob/master/LICENSE">GPL-v3 License</a>.<br />Running insert version here</p>
+		    </div>
+		</div>
+	    </div>
+	</div>
+    </body>
+    {% block endscripts %}{% endblock %}
+</html>
