+++ app/views/install/start.html
@@ -1,0 +1,61 @@
+{% extends "install/install_base.html" %}
+
+{% block content %}
+<div class="alert alert-info">For security reasons ensure it has a strong password and <strong>do not</strong> run this database under root credentials.</div>
+
+<p>Please fill out the database connection information that you will be using. This database user only needs the following permissions: <code>CREATE</code>, <code>INSERT</code>, <code>SELECT</code>, <code>UPDATE</code>, <code>DELETE</code>, <code>DROP</code>, and <code>ALTER</code>. This database user must be accessible from other servers.</p>
+
+<div class="alert alert-danger" id="warning" style="display: none;"></div>
+
+<fieldset>
+    <div class="form-group">
+	<label for="sql_db" class="control-label">Database Name</label>
+	<div>
+	    <input type="text" class="form-control" id="sql_db" autocomplete="off" required />
+	</div>
+    </div>
+    <div class="form-group">
+	<label for="sql_h" class="control-label">Database Host</label>
+	<div>
+	    <input type="text" class="form-control" id="sql_h" autocomplete="off" required />
+	</div>
+    </div>
+    <div class="form-group">
+	<label for="sql_u" class="control-label">Database User</label>
+	<div>
+	    <input type="text" class="form-control" id="sql_u" autocomplete="off" required />
+	</div>
+    </div>
+    <div class="form-group">
+	<label for="sql_p" class="control-label">Database User Password</label>
+	<div>
+	    <input type="password" class="form-control" id="sql_p" autocomplete="off" required />
+	</div>
+    </div>
+    <div class="form-group">
+	<div>
+	    <button id="submitBtn" class="btn btn-primary">Continue &rarr;</button>
+	</div>
+    </div>
+</fieldset>
+{% endblock %}
+
+{% block endscripts %}
+<script type="text/javascript">
+    $("#submitBtn").click(function () {
+	$.post("start", {
+	    sql_db: $("#sql_db").val(),
+	    sql_h: $("#sql_h").val(),
+	    sql_u: $("#sql_u").val(),
+	    sql_p: $("#sql_p").val()
+	}).done(function (data) {
+	    if (data === 'true') {
+
+	    } else {
+		$("#warning").text(data);
+		$("#warning").show();
+	    }
+	});
+    });
+</script>
+{% endblock %}
