+++ panel/router.php
@@ -38,24 +38,25 @@ $_SERVER['REMOTE_ADDR'] = (isset($_SERVER['HTTP_CF_CONNECTING_IP'])) ? $_SERVER[
 /*
 * Has Installer been run?
 */
-if(!file_exists(SRC_DIR.'core/configuration.php')) {
+if(!file_exists(SRC_DIR.'core/configuration.php') || (strpos($_SERVER['REQUEST_URI'], '/install') == 0 && !file_exists(SRC_DIR.'install.lock'))) {
 
 	if(file_exists(BASE_DIR.'vendor/autoload.php')) {
 
 		//pass off processing to the klein router for the installer
-		include(SRC_DIR.'core/routes/install/router.php');
-
+		include(SRC_DIR.'routes/install/router.php');
+		
 	} else {
 
 		//render the index page normally so that it shows the failures
 		include(PANEL_DIR.'install/install/index.php');
-
+		
 	}
-
+	
 	return;
 
 }
 
+require_once(SRC_DIR.'core/configuration.php');
 require_once(SRC_DIR.'core/autoloader.php');
 
 Twig_Autoloader::register();
@@ -83,8 +84,8 @@ ORM::configure(array(
 ));
 
 /*
-* Initalize Global
-*/
+ * Initalize Global core
+ */
 $core = new stdClass();
 $klein = new \Klein\Klein();
 
@@ -103,6 +104,18 @@ $core->twig = new Twig_Environment(new Twig_Loader_Filesystem(APP_DIR.'views/'),
 	'cache' => false,
 	'debug' => true
 ));
+
+/*
+ * Require HTTPS Connection
+ */
+if($core->settings->get('https') == 1) {
+
+	if(!$klein->request()->isSecure()) {
+		header("Location: https://".$klein->request()->server()['HTTP_HOST'].$klein->request()->server()['REQUEST_URI']);
+		return;
+	}
+	
+}
 
 /*
  * Check Language Settings
@@ -128,20 +141,6 @@ $core->twig->addGlobal('fversion', trim(file_get_contents(SRC_DIR.'versions/curr
 if($core->user->getData('root_admin') == 1) {
 	$core->twig->addGlobal('admin', true);
 }
-
-$klein->respond('/[*]', function($request, $response, $service, $app, $klein) use ($core) {
-
-	/*
-	* Require HTTPS Connection
-	*/
-	if($core->settings->get('https') == 1 && !$request->isSecure()) {
-
-		$response->redirect("https://".$request->server()['HTTP_HOST'].$request->server()['REQUEST_URI'])->send();
-		$klein->skipRemaining();
-
-	}
-
-});
 
 $klein->respond('!@^(/auth/|/langauge/|/api/)', function($request, $response, $service, $app, $klein) use ($core) {
