+++ src/core/autoloader.php
@@ -30,11 +30,12 @@ require_once(SRC_DIR.'core/components/errorhandler.php');
 require_once(SRC_DIR.'core/components/functions.php');
 require_once(SRC_DIR.'core/components/gsd.php');
 require_once(SRC_DIR.'core/components/page.php');
+require_once(SRC_DIR.'core/components/config.php');
+require_once(SRC_DIR.'core/components/url.php');
 
 /*
  * Include Required Global Class Files
  */
-require_once(SRC_DIR.'core/config.php');
 require_once(SRC_DIR.'core/authentication.php');
 require_once(SRC_DIR.'core/email.php');
 require_once(SRC_DIR.'core/files.php');
