+++ src/core/components/config.php
