--- src/core/routes/install/router.php
@@ -1,37 +1,0 @@
-<?php
-
-/*
-	PufferPanel - A Minecraft Server Management Panel
-	Copyright (c) 2015 PufferPanel
-
-	This program is free software: you can redistribute it and/or modify
-	it under the terms of the GNU General Public License as published by
-	the Free Software Foundation, either version 3 of the License, or
-	(at your option) any later version.
-
-	This program is distributed in the hope that it will be useful,
-	but WITHOUT ANY WARRANTY; without even the implied warranty of
-	MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-	GNU General Public License for more details.
-
-	You should have received a copy of the GNU General Public License
-	along with this program.  If not, see http://www.gnu.org/licenses/.
- */
-
-namespace PufferPanel\Core;
-
-require_once(BASE_DIR.'vendor/autoload.php');
-
-$klein = new \Klein\Klein();
-
-$klein->respond('/install?', function($request, $response, $service) {
-	include(PANEL_DIR.'install/install/index.php');
-});
-
-$klein->onHttpError(function($code, $klein) {
-	if($code == '404') {
-		$klein->response()->redirect('/install');
-	}
-});
-
-$klein->dispatch();
